==> czech/banksformer/nb0_create_dataset.ipynb <==
# Intro

This notebook creates preprocesses the raw czech data from multiple .csv files, which can be obtained here https://data.world/lpetrocelli/some-translatedreformatted-czech-banking-data  

The raw csv files should be placed in the 'raw_czech' directory, and in pariticular, you need to have the files 'new_transaction.csv', 'new_disposition.csv' and 'new_client.csv'. If you wish to put the files in a different location, change the RAW_CSV_DIR variable below to the location of the raw files.

import numpy as np
import pandas as pd
from matplotlib import pyplot as plt
import os

import pickle
from datetime import date

RAW_CSV_DIR = "raw_czech"

for fname in os.listdir(RAW_CSV_DIR):
    if ".csv" in fname and "new_" in fname:
        print(fname)
        csv_name = fname.split(".")[0]
        if RAW_CSV_DIR[-1] == "/":
            RAW_CSV_DIR = RAW_CSV_DIR[:-1]
        
        cmd = f"{csv_name} = pd.read_csv('{RAW_CSV_DIR}/{fname}')"
        print("Running:", cmd)
        exec(cmd)


### Sort by acct 

tr_by_acct = new_transaction.sort_values(by = ["account_id", "date"])

df = pd.merge(
        pd.merge(
            tr_by_acct,
            new_disposition[new_disposition["type"]=="OWNER"][["account_id", "client_id"]],
            on="account_id",
        ),
        new_client[["client_id", "age"]],
        on = "client_id"
)
df

from datetime import datetime
czech_date_parser = lambda x: datetime.strptime(str(x), "%y%m%d")

df["datetime"] = df["date"].apply(czech_date_parser)
df

df.to_csv("data/tr_by_acct_w_age.csv", index=False)
