==> czech/banksformer/nb1_preprocess_czech.ipynb <==
# Intro

Some basic preprocessing to ensure required features exist in the dataframe

import numpy as np
import pandas as pd
from matplotlib import pyplot as plt
import os

import pickle
from datetime import date

### Setup

ds_suffix = "-czech"
max_seq_len = 80
min_seq_len = 20

# Process

df = pd.read_csv("data/tr_by_acct_w_age.csv")
df

##### Sort by acct 

df = df.sort_values(by = ["account_id", "date"])

##### Date

from datetime import datetime
import calendar
czech_date_parser = lambda x: datetime.strptime(str(x), "%y%m%d")

df["datetime"] = df["date"].apply(czech_date_parser)

iso = df["datetime"].dt.isocalendar()

df["month"] = df["datetime"].dt.month
df["day"] = df["datetime"].dt.day
df["dow"] =  df["datetime"].dt.dayofweek
df["year"] = df["datetime"].dt.year

# dtme - days till month end
df["dtme"] = df.datetime.apply(lambda dt: calendar.monthrange(dt.year, dt.month)[1] - dt.day)
df

### Tcode

from field_config import cat_code_fields, TCODE_SEP

# create tcode by concating fields in "cat_code_fields"
def set_tcode(df, cat_code_fields):
    tcode = df[cat_code_fields[0]].astype(str)
    for ccf in cat_code_fields[1:]:
        tcode += TCODE_SEP + df[ccf].astype(str)

    df["tcode"] = tcode
    
    
set_tcode(df, cat_code_fields)

##### Time delta

df["td"] = df[["account_id", "datetime"]].groupby("account_id").diff()
df["td"] = df["td"].apply(lambda x: x.days)
df["td"].fillna(0.0, inplace=True)
df

# Write

folders = ['generated_data',
 'generated_data/parts',
 'stored_data',
 'generation_results',
 'data',
 'my_lib']


for f in folders:
    if not os.path.exists(f):
        os.mkdir(f)

df.to_csv(f"stored_data/final_df-{ds_suffix}.csv", index=False)
